chore: Implement bootstrappers for cluster, deployment tools, and resource management
- Added ClusterBootstrapper to handle cluster provisioning and validation.
- Introduced DeploymentToolBootstrapper for managing deployment tools and GitOps integration.
- Created GatewayControllerBootstrapper and IngressControllerBootstrapper for handling respective controllers.
- Implemented GitOpsSourceBootstrapper for OCI source registry management.
- Added MirrorRegistryBootstrapper for managing mirror registries in Docker environments.
- Introduced SecretManagerBootstrapper for SOPS secret management.
- Created IBootstrapper interface for consistent bootstrapper implementation.
- Added KubernetesResourceProvisionerExtensions for namespace management.
- Updated KSailUpCommandHandler to utilize new bootstrappers for improved modularity.
- Introduced ProjectCSIOption for CSI configuration in project options.
- Updated KSailClusterConfigLoader to support new options and configurations.
