+++ src/KSail.Models/DeploymentTool/KSailFluxDeploymentToolRepository.cs
@@ -8,7 +8,6 @@ namespace KSail.Models.DeploymentTool;
 /// </remarks>
 public class KSailFluxDeploymentToolRepository
 {
-
   [Description("The URL of the repository. [default: oci://ksail-registry:5000/ksail-registry]")]
   public Uri Url { get; set; } = new Uri("oci://ksail-registry:5000/ksail-registry");
 }
