+++ src/KSail.Models/Project/Enums/KSailCSIType.cs
@@ -1,0 +1,10 @@
+using System.ComponentModel;
+
+namespace KSail.Models.Project.Enums;
+
+
+public enum KSailCSIType
+{
+  [Description("Use the default CSI that comes with the chosen Kubernetes distribution.")]
+  Default
+}
